Use raw string for regex to avoid SyntaxWarning
in python 3 strings are considered unicode. the `\%` escape used in the regex is therefore treated as an unicode escape instead of a regex character escape. as this is not a valid unicode escape, python 3.12+ throws an SyntaxWarning.

treating the regex string as a raw string fixes the issue.

fixes https://github.com/sockless-coding/panasonic_cc/issues/206